test restart
